requirements-dev.txt
backcall==0.1.0
bleach==2.1.3
certifi==2018.4.16
coverage==4.5.1
decorator==4.3.0
Django==2.0.6
django-extensions==2.0.7
entrypoints==0.2.3
future==0.16.0
html5lib==1.0.1
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
model-mommy==1.6.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
pandocfilters==1.4.2
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.3
python-telegram-bot==10.1.0
pytz==2018.4
pyzmq==17.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
vidcutter==5.5.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.1
